extract toBase64DataUri util and fix hero image preview
